master add 64

+++ src/kernel/config_resolve.c
@@ -1,8 +1,8 @@
 #include "../ffi/FreeRTOSConfig.h"
 #include "../ffi/ffi.h"
 
-unsigned int CLINT_ADDR_=CLINT_ADDR;
-int CONFIG_TICK_RATE_HZ=configTICK_RATE_HZ;
-int CONFIG_CPU_CLOCK_HZ=configCPU_CLOCK_HZ;
-int CONFIG_ISR_STACK_SIZE_WORDS=configISR_STACK_SIZE_WORDS;
-int configMAX_PRIORITIES_=configMAX_PRIORITIES;
+unsigned long long CLINT_ADDR_=CLINT_ADDR;
+long long CONFIG_TICK_RATE_HZ=configTICK_RATE_HZ;
+long long CONFIG_CPU_CLOCK_HZ=configCPU_CLOCK_HZ;
+long long CONFIG_ISR_STACK_SIZE_WORDS=configISR_STACK_SIZE_WORDS;
+long long configMAX_PRIORITIES_=configMAX_PRIORITIES;
