+++ src/portable/portable.rs
@@ -80,7 +80,7 @@ pub fn v_port_setup_timer_interrupt() {
     let mut ul_current_time_high: UBaseType;
 
     unsafe {
-        pullNextTime = &ULL_NEXT_TIME as *const u64 as u32;
+        pullNextTime = &ULL_NEXT_TIME as *const u64 as usize;
         uxTimerIncrementsForOneTick = CONFIG_CPU_CLOCK_HZ / CONFIG_TICK_RATE_HZ;
         asm!("csrr {0}, mhartid",out(reg) ul_hart_id);
         pullMachineTimerCompareRegister = *ULL_MACHINE_TIMER_COMPARE_REGISTER_BASE + ul_hart_id * 4;
@@ -106,7 +106,7 @@ pub fn v_port_setup_timer_interrupt() {
 pub fn auto_set_currentTcb() {
     unsafe {
         match get_current_tcb() {
-            Some(x) => pxCurrentTCB = x as *const tskTaskControlBlock as u32,
+            Some(x) => pxCurrentTCB = x as *const tskTaskControlBlock as usize,
             None => pxCurrentTCB = 0,
         }
     }
@@ -115,10 +115,10 @@ pub fn auto_set_currentTcb() {
 /// Start up scheduler.
 pub fn x_port_start_scheduler() -> BaseType {
     unsafe {
-        xISRStackTop = &(X_ISRSTACK_.read()[CONFIG_ISR_STACK_SIZE_WORDS - 1]) as *const u32;
+        xISRStackTop = &(X_ISRSTACK_.read()[CONFIG_ISR_STACK_SIZE_WORDS - 1]) as *const usize;
     }
     v_port_setup_timer_interrupt();
-    let mut tmp: u32 = 0x800;
+    let mut tmp: usize = 0x800;
     if *CONFIG_MTIME_BASE_ADDRESS != 0 && *CONFIG_MTIMECMP_BASE_ADDRESS != 0 {
         tmp = 0x880;
     }
